Update README.md
I am removing talk of the --prefix option.  It will not work on a Mac.  It is assuming that you have checked out all the submodules independently and built and installed them with 'scons --prefix=...'.  i.e. the search path for includes/ now becomes the --prefix option and not the submodule directory.