rework submodules to avoid conflicts between source and installed
The raf source now uses a tool_raf.py file, so the raf tool in source is
found first if it exists, and any source modules which require 'raf' will
build against the source.  If the raf source is not in the tree, then they
will load the raf tool from eol_scons and build against the installed raf.

This depends on a fix in the eol_scons prefixoptions tool to a
long-standing problem with CPPPATH.  Now that OPT_PREFIX/include is always
added after CPPPATH, all the in-source include paths are searched first.
This avoids the problem where scons installs headers to the install prefix
so they can be found by other source modules.  Instead, all the local
header and library paths are searched first, so the source can be built
without installing anything first, and without building against installed
but outdated headers (problem zero for me).

This removes the use of --prefix to distinguish between a "source only"
build and a "use installed" build.  If a dependency exists in the source
tree, like vardb, raf, and so on, then the source in the tree will be used
for the build.  However, it should also work to build aircraft_oap against
the install prefix, as long as the submodules are empty.  --prefix can be
used to set both OPT_PREFIX and INSTALL_PREFIX, and it should work to both
build against and install into the same directory.  Any dependencies not in
the source tree will be looked for in the opt prefix.